+++ src/lib/flavour/routeHandler.ts
@@ -4,9 +4,7 @@ import { Flavour } from "types";
 import { FlavourLoader } from "./flavour";
 
 export function putFlavourHandler(req: Request, res: Response) {
-  const {
-    body: { flavour },
-  } = req ?? {};
+  const flavour = req?.body?.flavour;
 
   try {
     if (!flavour?.name?.length) throw "Flavour doesn't have a name";
@@ -15,22 +13,6 @@ export function putFlavourHandler(req: Request, res: Response) {
     saveFlavour(flavour as Flavour, true);
     res.status(200);
     return flavour;
-  } catch (e) {
-    res.status(500).json({ error: e });
-  }
-}
-
-export function postFlavourHandler(req: Request, res: Response) {
-  const {
-    body: { flavour },
-  } = req ?? {};
-
-  try {
-    if (!flavour?.name?.length) throw "Flavour doesn't have a name";
-    if (!flavour?.screens?.length) throw "Flavour doesn't have any screens";
-
-    saveFlavour(flavour as Flavour);
-    res.sendStatus(200);
   } catch (e) {
     res.status(500).json({ error: e });
   }